double of division sum

+++ src/loops/simple_programs_with_repitition/medium/q4/SerieslyQuestionMark.java
@@ -12,8 +12,8 @@ public class SerieslyQuestionMark {
         Scanner input = new Scanner(System.in);
         System.out.println("Enter a number for the upper bound of the series: ");
         int n = input.nextInt();
-        int sum = 0;
-        for (int i = 1; i < n; i++)
+        double sum = 0;
+        for (double i = 1; i < n; i++)
             sum += i / n;
         System.out.println("The sum of the series is " + sum);
     }
